deabstract locinfo and locdesc
